# packages/contracts/openapi/schema.yaml
openapi: 3.0.3
info:
  title: ECOnexão API
  version: 1.0.0
  description: API pública e administrativa versionada da ECOnexão.
paths:
...
  /api/v1/admin/routes/readiness:
    get:
      operationId: listAdminRouteReadiness
      description:
        DTO operacional, deliberadamente separado de contratos públicos
        de rota.
      parameters:
        - in: query
          name: region_slug
          schema:
            type: string
          required: true
      tags:
        - Admin routes
      security:
        - cookieAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegionRouteReadinessResponse'
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
          description: ''
        '401':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
          description: ''
        '403':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
          description: ''
        '429':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
          description: ''
        '500':
          content:
            application/json:
              schema:
                type: object
                properties:
                  detail:
                    type: string
          description: ''
...
  /api/v1/regions/{region_slug}/routes:
    get:
      operationId: listPublishedRegionRoutes
      summary: Listar rotas publicadas de uma região
      parameters:
        - in: path
          name: region_slug
          schema:
            type: string
          required: true
      tags:
        - Routes
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RouteSummary'
          description: ''
  /api/v1/regions/{region_slug}/routes/{route_slug}:
    get:
      operationId: retrievePublishedRoute
      summary: Obter uma rota publicada
      parameters:
        - in: path
          name: region_slug
          schema:
            type: string
          required: true
        - in: path
          name: route_slug
          schema:
            type: string
          required: true
      tags:
        - Routes
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RouteDetail'
          description: ''
  /api/v1/regions/{region_slug}/routes/{route_slug}/catalog:
    get:
      operationId: listPublishedRouteCatalog
      description:
        Retorna somente atores publicados, localizações visíveis na região
        e contatos autorizados para publicação.
      summary: Listar o catálogo público contextual da rota
      parameters:
        - in: path
          name: region_slug
          schema:
            type: string
          required: true
        - in: path
          name: route_slug
          schema:
            type: string
          required: true
      tags:
        - Catalog
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RouteCatalogItem'
          description: ''
...
components:
  schemas:
...
    Alert:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        severity:
          enum:
            - info
            - warning
            - critical
          type: string
          description: |-
            * `info` - Informativo
            * `warning` - Atenção
            * `critical` - Crítico
          x-spec-enum-id: f57d55d8feea7d7d
        title:
          type: string
          maxLength: 160
        description:
          type: string
        alternative:
          type: string
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
          nullable: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
        - description
        - id
        - severity
        - starts_at
        - title
        - updated_at
...
    OperatingHours:
      type: object
      properties:
        weekday:
          type: integer
          maximum: 32767
          minimum: 0
          nullable: true
        opens_at:
          type: string
          format: time
          nullable: true
        closes_at:
          type: string
          format: time
          nullable: true
        is_closed:
          type: boolean
        exception_date:
          type: string
          format: date
          nullable: true
        public_note:
          type: string
          maxLength: 240
...
    PublicActor:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        slug:
          type: string
          maxLength: 140
          pattern: ^[-a-zA-Z0-9_]+$
        public_name:
          type: string
          maxLength: 160
        actor_kind:
          enum:
            - business
            - individual_provider
            - community
            - institution
            - support
          type: string
          description: |-
            * `business` - Empresa
            * `individual_provider` - Prestador individual
            * `community` - Comunidade
            * `institution` - Instituição
            * `support` - Ponto de apoio
          x-spec-enum-id: 697952b537e1f241
        short_description:
          type: string
          maxLength: 180
        full_description:
          type: string
        services: {}
        partnership_type:
          enum:
            - editorial
            - partner
            - sponsored
          type: string
          description: |-
            * `editorial` - Editorial
            * `partner` - Parceiro
            * `sponsored` - Patrocinado
          x-spec-enum-id: 38e3f2ee39ffa790
        category_slug:
          type: string
          readOnly: true
          pattern: ^[-a-zA-Z0-9_]+$
        category_name:
          type: string
          readOnly: true
        locations:
          type: array
          items:
            $ref: '#/components/schemas/PublicActorLocation'
          readOnly: true
        contact_channels:
          type: array
          items:
            $ref: '#/components/schemas/PublicContactChannel'
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
        - actor_kind
        - category_name
        - category_slug
        - contact_channels
        - id
        - locations
        - public_name
        - short_description
        - slug
        - updated_at
    PublicActorLocation:
      type: object
      properties:
        label:
          type: string
          maxLength: 120
        address_fields: {}
        point:
          type: object
          required:
            - type
            - coordinates
          properties:
            type:
              type: string
              enum:
                - Point
            coordinates:
              type: array
              items:
                type: number
                format: double
              minItems: 2
          readOnly: true
          nullable: true
        is_primary:
          type: boolean
        operating_hours:
          type: array
          items:
            $ref: '#/components/schemas/OperatingHours'
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
        - label
        - operating_hours
        - point
        - updated_at
    PublicContactChannel:
      type: object
      properties:
        channel_type:
          enum:
            - phone
            - whatsapp
            - email
            - website
            - instagram
          type: string
          description: |-
            * `phone` - Telefone
            * `whatsapp` - WhatsApp
            * `email` - E-mail
            * `website` - Site
            * `instagram` - Instagram
          x-spec-enum-id: f040ddfdb8eedb9d
        public_value:
          type: string
          maxLength: 500
        verified_at:
          type: string
          format: date-time
          nullable: true
      required:
        - channel_type
...
    ReadinessDimensions:
      type: object
      properties:
        content:
          type: integer
          maximum: 100
          minimum: 0
        trace:
          type: integer
          maximum: 100
          minimum: 0
        catalog:
          type: integer
          maximum: 100
          minimum: 0
        alerts:
          type: integer
          maximum: 100
          minimum: 0
        offline:
          type: integer
          maximum: 100
          minimum: 0
      required:
        - alerts
        - catalog
        - content
        - offline
        - trace
    RegionRouteReadinessResponse:
      type: object
      properties:
        region_slug:
          type: string
        routes:
          type: array
          items:
            $ref: '#/components/schemas/RouteReadiness'
      required:
        - region_slug
        - routes
...
    RouteCatalogItem:
      type: object
      properties:
        route_role:
          enum:
            - experience
            - support
            - start
            - stop
            - emergency
            - service
          type: string
          description: |-
            * `experience` - Experiência
            * `support` - Apoio
            * `start` - Início
            * `stop` - Parada
            * `emergency` - Emergência
            * `service` - Serviço
          x-spec-enum-id: 26029888495bbdd4
        editorial_position:
          type: integer
          maximum: 32767
          minimum: 0
        is_featured:
          type: boolean
        sponsorship_label:
          type: string
          maxLength: 160
        stage_id:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        actor:
          allOf:
            - $ref: '#/components/schemas/PublicActor'
          readOnly: true
      required:
        - actor
        - route_role
        - stage_id
    RouteDetail:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        slug:
          type: string
          maxLength: 120
          pattern: ^[-a-zA-Z0-9_]+$
        public_name:
          type: string
          maxLength: 160
        short_promise:
          type: string
          maxLength: 180
        duration_minutes:
          type: integer
          maximum: 2147483647
          minimum: 0
        difficulty:
          enum:
            - easy
            - moderate
            - hard
          type: string
          description: |-
            * `easy` - Fácil
            * `moderate` - Moderada
            * `hard` - Difícil
          x-spec-enum-id: d835a0185ead1283
        estimated_cost_min:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          nullable: true
        estimated_cost_max:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          nullable: true
        transport_modes: {}
        accessibility_content:
          type: string
        offline_enabled:
          type: boolean
        updated_at:
          type: string
          format: date-time
          readOnly: true
        region_slug:
          type: string
          readOnly: true
          pattern: ^[-a-zA-Z0-9_]+$
        region_name:
          type: string
          readOnly: true
        description:
          type: string
        preparation_content:
          type: string
        stages:
          type: array
          items:
            $ref: '#/components/schemas/RouteStage'
          readOnly: true
        segments:
          type: array
          items:
            $ref: '#/components/schemas/RouteSegment'
          readOnly: true
        alerts:
          type: array
          items:
            $ref: '#/components/schemas/Alert'
          readOnly: true
      required:
        - alerts
        - difficulty
        - duration_minutes
        - id
        - public_name
        - region_name
        - region_slug
        - segments
        - short_promise
        - slug
        - stages
        - updated_at
    RouteReadiness:
      type: object
      properties:
        route_id:
          type: string
          format: uuid
        slug:
          type: string
        title:
          type: string
        editorial_status:
          enum:
            - draft
            - review
            - approved
            - published
            - suspended
            - archived
          type: string
          description: |-
            * `draft` - Rascunho
            * `review` - Em revisão
            * `approved` - Aprovado
            * `published` - Publicado
            * `suspended` - Suspenso
            * `archived` - Arquivado
          x-spec-enum-id: b8cd74618bec600b
        formula_version:
          type: string
        weights:
          $ref: '#/components/schemas/ReadinessDimensions'
        dimensions:
          $ref: '#/components/schemas/ReadinessDimensions'
        score:
          type: integer
          maximum: 100
          minimum: 0
          nullable: true
        is_ready:
          type: boolean
        blocking_reasons:
          type: array
          items:
            type: string
        missing_required_fields:
          type: array
          items:
            type: string
        stages_count:
          type: integer
          minimum: 0
        segments_count:
          type: integer
          minimum: 0
        published_points_count:
          type: integer
          minimum: 0
        points_in_review_count:
          type: integer
          minimum: 0
        verified_contacts_count:
          type: integer
          minimum: 0
        unverified_public_contacts_count:
          type: integer
          minimum: 0
        blocking_alerts_count:
          type: integer
          minimum: 0
        last_revision_at:
          type: string
          format: date-time
          nullable: true
        published_version:
          type: integer
          minimum: 1
          nullable: true
      required:
        - blocking_alerts_count
        - blocking_reasons
        - dimensions
        - editorial_status
        - formula_version
        - is_ready
        - last_revision_at
        - missing_required_fields
        - points_in_review_count
        - published_points_count
        - published_version
        - route_id
        - score
        - segments_count
        - slug
        - stages_count
        - title
        - unverified_public_contacts_count
        - verified_contacts_count
        - weights
    RouteSegment:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        from_stage_id:
          type: string
          format: uuid
          readOnly: true
        to_stage_id:
          type: string
          format: uuid
          readOnly: true
        geometry:
          type: object
          required:
            - type
            - coordinates
          properties:
            type:
              type: string
              enum:
                - LineString
            coordinates:
              type: array
              items:
                type: array
                items:
                  type: number
                  format: double
                minItems: 2
              minItems: 2
          readOnly: true
        transport_mode:
          type: string
          maxLength: 32
        distance_meters:
          type: integer
          maximum: 2147483647
          minimum: 0
        duration_minutes:
          type: integer
          maximum: 2147483647
          minimum: 0
        instructions:
          type: string
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
        - distance_meters
        - duration_minutes
        - from_stage_id
        - geometry
        - id
        - to_stage_id
        - transport_mode
        - updated_at
    RouteStage:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        position:
          type: integer
          maximum: 32767
          minimum: 0
        public_name:
          type: string
          maxLength: 160
        description:
          type: string
        point:
          type: object
          required:
            - type
            - coordinates
          properties:
            type:
              type: string
              enum:
                - Point
            coordinates:
              type: array
              items:
                type: number
                format: double
              minItems: 2
          readOnly: true
        arrival_guidance:
          type: string
        duration_minutes:
          type: integer
          maximum: 2147483647
          minimum: 0
          nullable: true
        stage_type:
          enum:
            - start
            - stop
            - experience
            - support
            - end
          type: string
          description: |-
            * `start` - Início
            * `stop` - Parada
            * `experience` - Experiência
            * `support` - Apoio
            * `end` - Fim
          x-spec-enum-id: 01800cd6cab71c47
        is_optional:
          type: boolean
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
        - id
        - point
        - position
        - public_name
        - stage_type
        - updated_at
    RouteSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        slug:
          type: string
          maxLength: 120
          pattern: ^[-a-zA-Z0-9_]+$
        public_name:
          type: string
          maxLength: 160
        short_promise:
          type: string
          maxLength: 180
        duration_minutes:
          type: integer
          maximum: 2147483647
          minimum: 0
        difficulty:
          enum:
            - easy
            - moderate
            - hard
          type: string
          description: |-
            * `easy` - Fácil
            * `moderate` - Moderada
            * `hard` - Difícil
          x-spec-enum-id: d835a0185ead1283
        estimated_cost_min:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          nullable: true
        estimated_cost_max:
          type: string
          format: decimal
          pattern: ^-?\d{0,8}(?:\.\d{0,2})?$
          nullable: true
        transport_modes: {}
        accessibility_content:
          type: string
        offline_enabled:
          type: boolean
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
        - difficulty
        - duration_minutes
        - id
        - public_name
        - short_promise
        - slug
        - updated_at
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: econexao_admin_sessionid